requirements.txt
streamlit==1.34.0
google-generativeai==0.8.4
tqdm==4.66.2
protobuf==4.25.3
google-api-core==2.18.0
google-auth==2.29.0
python-dotenv==1.1.0